+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Mnemonics docs:
   - fix `BRA` reference docs
   - fix duplicate entry in instruction docs for `BCHG`, `BCLR`, `BSET`
+- Parser: correct parsing for `ADDQ.B`/`SUBQ.B`
 
 ## 0.2.0 - 2025-10-31